Chore: clean up code
