+++ Gherkin/common.py
@@ -1,11 +1,7 @@
-from os import getenv
 import re
 import chromadb
-from chromadb import Documents, EmbeddingFunction, Embeddings
 from chromadb.config import Settings
-from chromadb.utils.embedding_functions import SentenceTransformerEmbeddingFunction
-import requests
-import json
+import common_embed
 
 ### parse model@host
 
@@ -144,84 +140,6 @@ def get_internal_id_of_keyword(internal_description_id: str) -> str:
 
 ### database queries
 
-def call_server(url: str, token: str, payload: dict[str, str]) -> dict[str, str]:
-
-    headers = {"Authorization": f"Bearer {token}"}
-
-    try:
-        response = requests.post(url, headers=headers, json=payload)
-        response.raise_for_status()
-    except requests.exceptions.HTTPError as e:
-        error_message = response.text
-        print(url + "\nreturned error\n" + error_message, flush=True)
-        raise Exception("Model failed") from e
-    except Exception as e:
-        print(url + "\nfailed with exception\n" + str(e), flush=True)
-        raise Exception("An error occurred") from e
-    txt = response.text
-    jsonz = {}
-    try:
-        jsonz = json.loads(txt)
-    except BaseException as e:
-        print(f"Error while trying to extract JSON\n{e}\nThe API answer is\n{txt}", flush=True)
-        raise Exception(f"Error while trying to extract JSON from \"{txt}\", problem is: " + str(e)) from e
-    return jsonz
-
-class TogetherEmbeddingFunction(EmbeddingFunction[Documents]):
-    def __init__(self, model: str):
-        self.model = model
-
-    def __call__(self, input: Documents) -> Embeddings:
-        # see https://docs.together.ai/docs/embeddings-overview#generating-multiple-embeddings
-        url = "https://api.together.xyz/v1/embeddings"
-        token = getenv("TOGETHER_API_KEY")
-        payload = {
-             "model": self.model,
-             "input": input
-            }
-
-        result = call_server(url, token, payload)
-        return [d['embedding'] for d in result['data']]
-
-class MistralEmbeddingFunction(EmbeddingFunction[Documents]):
-    def __init__(self, model: str):
-        self.model = model
-
-    def __call__(self, input: Documents) -> Embeddings:
-        url = "https://api.mistral.ai/v1/embeddings"
-        token = getenv("MISTRAL_API_KEY")
-        payload = {
-             "model": self.model,
-             "input": input
-            }
-
-        result = call_server(url, token, payload)
-        return [d['embedding'] for d in result['data']]
-
-class HuggingFaceEmbeddingFunction(EmbeddingFunction[Documents]):
-    def __init__(self, model: str):
-        self.model = model
-
-    def __call__(self, input: Documents) -> Embeddings:
-        url = f"https://api-inference.huggingface.co/pipeline/feature-extraction/{self.model}"
-        token = getenv("HUGGINGFACE_API_KEY")
-        payload = {
-             "inputs": input
-            }
-
-        result = call_server(url, token, payload)
-        return result
-
-def build_embedding_function(host: str, model: str) -> SentenceTransformerEmbeddingFunction:
-    if host == None:
-        return SentenceTransformerEmbeddingFunction(model_name=model)
-    if host == "Together":
-        return TogetherEmbeddingFunction(model)
-    if host == "Mistral":
-        return MistralEmbeddingFunction(model)
-    if host == "HuggingFace":
-        return HuggingFaceEmbeddingFunction(model)
-    raise ValueError(f"Unknown host ({host})")
 
 
 def search_keywords(db_path: str, host: str, model: str, project: str, keyword_type: str, keyword: str, nb_results:int) -> list[dict[str, str]]:
@@ -255,7 +173,7 @@ def search_keywords(db_path: str, host: str, model: str, project: str, keyword_t
     # Get the appropriate collection
     collection_name = f"{get_collection_name(model, host, project, keyword_type)}"
 
-    embedding_function = build_embedding_function(host, model)
+    embedding_function = common_embed.build_embedding_function(host, model)
     try:
         collection = chroma_client.get_collection(
             name=collection_name,
